+++ eagle-core/eagle-alert-parent/eagle-alert/alert-engine/src/main/java/org/apache/eagle/alert/engine/interpreter/PolicyValidationResult.java
@@ -14,18 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.service.metadata.resource;
+package org.apache.eagle.alert.engine.interpreter;
+
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import org.apache.commons.lang3.exception.ExceptionUtils;
+import org.apache.eagle.alert.engine.coordinator.PolicyDefinition;
 
 @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
-public class PolicyParseResult {
+public class PolicyValidationResult {
     private boolean success;
     private String message;
     private String exception;
 
     private PolicyExecutionPlan policyExecutionPlan;
+    private PolicyDefinition policyDefinition;
 
     public String getException() {
         return exception;
@@ -61,5 +64,13 @@ public class PolicyParseResult {
 
     public void setPolicyExecutionPlan(PolicyExecutionPlan policyExecutionPlan) {
         this.policyExecutionPlan = policyExecutionPlan;
+    }
+
+    public PolicyDefinition getPolicyDefinition() {
+        return policyDefinition;
+    }
+
+    public void setPolicyDefinition(PolicyDefinition policyDefinition) {
+        this.policyDefinition = policyDefinition;
     }
 }
